# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Gallery API",
        "description": "Internal HTTP API for the self-hosted media gallery — a **hybrid REST-RPC**\ninterface served under the `/api` base path. All responses are JSON.\n\n**Conventions**\n- Reads return `200` with the resource or collection.\n- Creates return `201` with the created resource.\n- Updates (`PUT`/`PATCH`) return `200` with the updated resource.\n- Deletes return `204` with no body.\n- Non-CRUD actions (login, migrate, scan, bulk operations) are `POST`\n  verbs that return `200` with a small result object.\n- Errors return `{ \"error\": \"<MachineCode>\", \"message\": \"<human text>\" }`\n  with a `4xx`/`5xx` status. `error` is a stable PascalCase code safe to\n  branch on; `message` is for display and may change.\n\n**Auth** — a Bearer token (`Authorization: Bearer <token>`) obtained from\n`POST /auth/login` (valid 24h). All `GET`s are public; state-changing\nmethods require a token, except a small public allowlist (media tagging\nand the batched `POST /media/by-ids` read).",
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Same-origin API base path"
        }
    ],
    "paths": {
...
        "/tag-implications": {
            "get": {
                "tags": [
                    "Tag Implications"
                ],
                "summary": "List tag implications",
                "description": "GET /tag-implications — All tag implications. Cached.",
                "operationId": "a0ce21a9789fd654d9f7b164491bc60b",
                "responses": {
                    "200": {
                        "description": "All implications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TagImplication"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tag Implications"
                ],
                "summary": "Create a tag implication",
                "description": "POST /tag-implications — Add \"tag_id implies implied_tag_id\".\nBody: { tag_id, implied_tag_id }. Rejects cycles and self-implication.\nReturns the created implication.",
                "operationId": "9ed70fe6fc2fe1fdc19be3a497665936",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tag_id",
                                    "implied_tag_id"
                                ],
                                "properties": {
                                    "tag_id": {
                                        "type": "integer"
                                    },
                                    "implied_tag_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The created implication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagImplication"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "CannotImplySelf / CycleDetected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TagDoesNotExist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tag-implications/{tag_id}/{implied_tag_id}": {
            "delete": {
                "tags": [
                    "Tag Implications"
                ],
                "summary": "Remove a tag implication",
                "description": "DELETE /tag-implications/{tag_id}/{implied_tag_id} — Remove an implication.",
                "operationId": "381181baeba3505869a9c5d513610990",
                "parameters": [
                    {
                        "name": "tag_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "implied_tag_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed"
                    },
                    "400": {
                        "description": "InvalidTagID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "description": "Standard error envelope.",
                "properties": {
                    "error": {
                        "description": "Stable PascalCase machine code (e.g. MediaNotFound).",
                        "type": "string"
                    },
                    "message": {
                        "description": "Human-readable message for display.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TagImplication": {
                "description": "A tag implication (tag -> implied tag).",
                "properties": {
                    "tag_id": {
                        "type": "integer"
                    },
                    "tag_name": {
                        "type": "string"
                    },
                    "implied_tag_id": {
                        "type": "integer"
                    },
                    "implied_tag_name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Admin bearer token issued by POST /auth/login (valid 24 hours).",
                "scheme": "bearer"
            }
        }
    },
...
}